injected wallet interface renamed to wallet interface and moved to a separate file
