+++ lib/wallet/index.d.ts
@@ -1,4 +1,5 @@
-import { WalletConnection, RequestSignTransactionsOptions, SignInOptions } from './wallet-connection';
+import { Wallet, RequestSignTransactionsOptions, SignInOptions } from './wallet-interface';
+import { WalletConnection } from './wallet-connection';
 import { WalletConnectionWithKeyManagement } from './wallet-connection-with-key-management';
 import { WalletConnectionInjected } from './injected/wallet-connection-injected';
 import { WalletConnectionSender } from './injected/wallet-connection-sender';
@@ -6,5 +7,5 @@ import { WalletConnectionInjectedWithFckManagement } from './injected/wallet-con
 import { WalletConnectionRedirect, ConnectedWalletAccountRedirect } from './redirect/wallet-connection-redirect';
 import { createWalletConnection, WalletConnectionType, WalletConnectionParameterOptions } from './wallet-connection-creator';
 import * as wallet_injected_interface from './injected/wallet-injected-interface';
-export { createWalletConnection, WalletConnectionType, WalletConnectionParameterOptions, WalletConnection, RequestSignTransactionsOptions, SignInOptions, WalletConnectionWithKeyManagement, WalletConnectionInjected, WalletConnectionSender, WalletConnectionInjectedWithFckManagement, WalletConnectionRedirect, ConnectedWalletAccountRedirect, //TODO: do we need this export?
+export { Wallet, RequestSignTransactionsOptions, SignInOptions, createWalletConnection, WalletConnectionType, WalletConnectionParameterOptions, WalletConnection, WalletConnectionWithKeyManagement, WalletConnectionInjected, WalletConnectionSender, WalletConnectionInjectedWithFckManagement, WalletConnectionRedirect, ConnectedWalletAccountRedirect, //TODO: do we need this export?
 wallet_injected_interface, };
