+++ lib/wallet/injected/wallet-connection-injected-with-fck-management.d.ts
@@ -1,4 +1,4 @@
-import { RequestSignTransactionsOptions, SignInOptions } from "../wallet-connection";
+import { RequestSignTransactionsOptions, SignInOptions } from "../wallet-interface";
 import { WalletConnectionWithKeyManagement, ConnectedWalletAccountWithKeyManagement } from "../wallet-connection-with-key-management";
 import { Near } from '../../near';
 import { Account } from '../../account';
