+++ lib/wallet/wallet-connection-creator.js
@@ -18,7 +18,7 @@ function createWalletConnection(near, appKeyPrefix, { type, data }) {
             return new index_1.WalletConnectionInjected(near, appKeyPrefix, data.walletName);
         }
         case WalletConnectionType.INJECTED_SENDER_WALLET: {
-            console.warn("Temporary implementation, should be relaced with WalletConnectionInjected");
+            console.warn('Temporary implementation, should be relaced with WalletConnectionInjected');
             return new index_1.WalletConnectionSender(near, appKeyPrefix, data.walletName);
         }
         case WalletConnectionType.INJECTED_WITH_FCK_MANAGEMENT: {
