+++ lib/wallet/wallet-connection.d.ts
@@ -1,7 +1,7 @@
 import { Account } from '../account';
 import { Connection } from '../connection';
 import { Near } from '../near';
-import { Transaction } from '../transaction';
+import { RequestSignTransactionsOptions, SignInOptions } from './wallet-interface';
 export declare abstract class WalletConnection implements TransactionsSigner, SignInProvider, AcocuntProvider {
     /** @hidden */
     _near: Near;
@@ -25,23 +25,6 @@ export declare abstract class WalletConnection implements TransactionsSigner, Si
 export declare abstract class ConnectedWalletAccount extends Account {
     walletConnection: WalletConnection;
     constructor(walletConnection: WalletConnection, connection: Connection, accountId: string);
-}
-/**
- * Information to send NEAR wallet for signing transactions and redirecting the browser back to the calling application
- */
-export interface RequestSignTransactionsOptions {
-    /** list of transactions to sign */
-    transactions: Transaction[];
-    /** url NEAR Wallet will redirect to after transaction signing is complete */
-    callbackUrl?: string;
-    /** meta information NEAR Wallet will send back to the application. `meta` will be attached to the `callbackUrl` as a url search param */
-    meta?: string;
-}
-export interface SignInOptions {
-    contractId?: string;
-    methodNames?: string[];
-    successUrl?: string;
-    failureUrl?: string;
 }
 interface TransactionsSigner {
     requestSignTransactions({ transactions, meta, callbackUrl }: RequestSignTransactionsOptions): Promise<void>;
